Adds secondary frame buffer for LED display.
Otherwise, the buffer may be drawn or redrawn out of sync with the intended refresh.
That is, when the display is triggered in code, the buffer is not read and drawn
until some time after. Here, when the display is triggered, this just swaps the buffers
by copying the data from the first buffer into a second. Then, the draw procedure just
looks at the second buffer.

This allows for the first buffer to be written to and erased out of sync with the
update. This isn't real hardware... and there is no way to keep code in sync with
the actual display refresh since it is arbitrary and platform dependent (yikes.)

We could simply swap the buffer pointer, but that requires changes to how code to
set and get LED display data works... it would have to check for which buffer is
currently active.

What can you do.